inventory

+++ app/Http/Controllers/AgentVouchersController.php
@@ -180,6 +180,9 @@ class AgentVouchersController extends Controller
     public function show($vid)
     {
 		$voucher = Voucher::where('id',$vid)->where('agent_id',Auth::user()->id)->first();
+		if (empty($voucher)) {
+            return abort(404); //record not found
+        }
 		if($voucher->status_main  > 3)
 		{
 			return redirect()->route('agentVoucherView',$voucher->id);
@@ -200,6 +203,10 @@ class AgentVouchersController extends Controller
     public function edit($id)
     {
         $record = Voucher::where('id',$id)->where('agent_id',Auth::user()->id)->first();
+		if (empty($record)) {
+            return abort(404); //record not found
+        }
+		
 		$countries = Country::where('status', 1)->orderBy('name', 'ASC')->get();
 		$airlines = Airline::where('status', 1)->orderBy('name', 'ASC')->get();
 		$customer = Customer::where('id',$record->customer_id)->first();
@@ -296,6 +303,9 @@ class AgentVouchersController extends Controller
     public function destroy($id)
     {
         $record = Voucher::where('id',$id)->where('agent_id',Auth::user()->id)->first();
+		if (empty($record)) {
+            return abort(404); //record not found
+        }
 		//$voucherHotel = VoucherHotel::where('voucher_id',$id)->delete();
 		$voucherActivity = VoucherActivity::where('voucher_id',$id)->delete();
 		
@@ -515,123 +525,9 @@ class AgentVouchersController extends Controller
 	public function destroyActivityFromVoucher($id)
     {
         $record = VoucherActivity::find($id);
+		
         $record->delete();
         return redirect()->back()->with('success', 'Activity Deleted Successfully.');
-    }
-	
- public function voucherActivityItineraryPdf(Request $request, $vid)
-    {
-		$voucher = Voucher::where('id',$vid)->where('agent_id',Auth::user()->id)->first();
-		if (empty($voucher)) {
-            return abort(404); //record not found
-        }
-		$voucherHotel = VoucherHotel::where('voucher_id',$voucher->id)->get();
-		$voucherActivity = VoucherActivity::where('voucher_id',$voucher->id)->get();
-       
-        
-        
-
-        $pdf = SPDF::loadView('vouchers.ActivityItineraryPdf', compact('voucher','voucherHotel','voucherActivity'));
-       $pdf->setPaper('A4')->setOrientation('portrait');
-        return $pdf->download('Itinerary'.$vid.'.pdf');
-		
-	
-	
-	return \Response::make($content,200, $headers);
-    }
-	
-	public function voucherInvoicePdf(Request $request, $vid)
-    {
-		$voucher = Voucher::where('id',$vid)->where('agent_id',Auth::user()->id)->first();
-		if (empty($voucher)) {
-            return abort(404); //record not found
-        }
-		
-		$voucherHotel = VoucherHotel::where('voucher_id',$voucher->id)->get();
-		$agent = User::where('id',$voucher->agent_id)->first();
-		$customer = Customer::where('id',$voucher->customer_id)->first();
-		
-       $voucherActivity = VoucherActivity::where('voucher_id',$voucher->id)->get();
-	  
-		
-        $dataArray = [];
-		$discountTotal = 0;
-		$subTotal = 0;
-		
-		if(!empty($voucherActivity)){
-					 foreach($voucherActivity as $kkh => $ap)
-					 {
-						
-					$activity = SiteHelpers::getActivity($ap->activity_id);
-					$vat =  1 + $activity->vat;
-					$vatPrice = $ap->totalprice/$vat;
-					//$total = ($ap->totalprice+$ap->discountPrice) - ($vatPrice);
-					$total = $ap->totalprice;
-				$dataArray[$ap->variant_code.$kkh] = [
-				'hhotelActName' => $activity->title.'-'.$ap->variant_name,
-				'TouCheckInCheckOutDate' => $ap->tour_date,
-				'adult' => $ap->adult,
-				'child' => $ap->child,
-				'NoofPax' => 0,
-				'hotel' => 0,
-				'totalprice' => $total,
-				];
-				$discountTotal += $ap->discountPrice;
-				$subTotal+= $total ;
-					 }
-					
-			}
-			
-		if(!empty($voucherHotel)){
-					 foreach($voucherHotel as $kk => $vh)
-					 {
-					  $hotelData = json_decode($vh->hotel_other_details);
-					  $noofPax = 0;
-						$netRate = 0;
-					  foreach($hotelData as $k => $hd)
-					  {
-						  
-						  $noofPax+= $hd->nop_s;
-						  $noofPax+= $hd->nop_d;
-						  $noofPax+= $hd->nop_eb;
-						  $noofPax+= $hd->nop_cwb;
-						  $noofPax+= $hd->nop_cnb;
-						  
-						  $netRate+= $hd->nr_s;
-						  $netRate+= $hd->nr_d;
-						  $netRate+= $hd->nr_eb;
-						  $netRate+= $hd->nr_cwb;
-						  $netRate+= $hd->nr_cnb;
-					  }
-				$dataArray[$vh->hotel->name.$kk] = [
-				'hhotelActName' => $vh->hotel->name,
-				'TouCheckInCheckOutDate' =>$vh->check_in_date.' / '.$vh->check_out_date,
-				'NoofPax' => $noofPax,
-				'adult' => 0,
-				'child' => 0,
-				'hotel' => 1,
-				'totalprice' => $netRate,
-				];
-				$subTotal+= $netRate;
-					 }
-					
-			}
-			$withVatTotalAmount = $subTotal;
-			$vatGrand = '1.05';
-			$grandWithOutVatTotal = ($withVatTotalAmount/$vatGrand) - $discountTotal;
-			$grandWithVatTotal = ($withVatTotalAmount) - $discountTotal;
-			$vatTotal = ($withVatTotalAmount) - $grandWithOutVatTotal;
-			
-			
-       
-//return view('vouchers.invoicePdf', compact('dataArray','agent','customer','voucher','discountTotal','grandWithVatTotal','grandWithOutVatTotal','vatTotal'));
-        $pdf = SPDF::loadView('vouchers.invoicePdf', compact('dataArray','agent','customer','voucher','discountTotal','grandWithOutVatTotal','grandWithVatTotal','vatTotal'));
-       $pdf->setPaper('A4')->setOrientation('portrait');
-        return $pdf->download('Invoice'.$vid.'.pdf');
-		
-	
-	
-	return \Response::make($content,200, $headers);
     }
 	
 
@@ -737,6 +633,9 @@ class AgentVouchersController extends Controller
 	 public function agentVoucherView($vid)
     {
 		$voucher =  Voucher::where('id',$vid)->where('agent_id',Auth::user()->id)->first();
+		if (empty($voucher)) {
+            return abort(404); //record not found
+        }
 		$voucherActivity = VoucherActivity::where('voucher_id',$voucher->id)->get();
 	
 		$voucherStatus = config("constants.voucherStatus");
